separate diff tests + check psr for sequences
