+++ tests/test_circuit.py
@@ -7,134 +7,15 @@ import torch
 from torch import Tensor
 
 import pyqtorch as pyq
-from pyqtorch import DiffMode, expectation, run, sample
+from pyqtorch import run, sample
 from pyqtorch.circuit import QuantumCircuit
-from pyqtorch.matrices import COMPLEX_TO_REAL_DTYPES
 from pyqtorch.noise import Noise
 from pyqtorch.parametric import Parametric
 from pyqtorch.primitive import Primitive
 from pyqtorch.utils import (
-    GRADCHECK_ATOL,
     DensityMatrix,
     product_state,
 )
-
-
-# TODO add GPSR when multigap is implemented for this test
-@pytest.mark.parametrize("n_qubits", [3, 4, 5])
-def test_adjoint_diff(n_qubits: int) -> None:
-    rx = pyq.RX(0, param_name="theta_0")
-    cry = pyq.CPHASE(0, 1, param_name="theta_1")
-    rz = pyq.RZ(2, param_name="theta_2")
-    cnot = pyq.CNOT(1, 2)
-    ops = [rx, cry, rz, cnot]
-    circ = pyq.QuantumCircuit(n_qubits, ops)
-    obs = pyq.QuantumCircuit(n_qubits, [pyq.Z(0)])
-
-    theta_0_value = torch.pi / 2
-    theta_1_value = torch.pi
-    theta_2_value = torch.pi / 4
-
-    state = pyq.zero_state(n_qubits)
-
-    theta_0_ad = torch.tensor([theta_0_value], requires_grad=True)
-    thetas_0_adjoint = torch.tensor([theta_0_value], requires_grad=True)
-
-    theta_1_ad = torch.tensor([theta_1_value], requires_grad=True)
-    thetas_1_adjoint = torch.tensor([theta_1_value], requires_grad=True)
-
-    theta_2_ad = torch.tensor([theta_2_value], requires_grad=True)
-    thetas_2_adjoint = torch.tensor([theta_2_value], requires_grad=True)
-
-    values_ad = {"theta_0": theta_0_ad, "theta_1": theta_1_ad, "theta_2": theta_2_ad}
-    values_adjoint = {
-        "theta_0": thetas_0_adjoint,
-        "theta_1": thetas_1_adjoint,
-        "theta_2": thetas_2_adjoint,
-    }
-    exp_ad = expectation(circ, state, values_ad, obs, DiffMode.AD)
-    exp_adjoint = expectation(circ, state, values_adjoint, obs, DiffMode.ADJOINT)
-
-    grad_ad = torch.autograd.grad(
-        exp_ad, tuple(values_ad.values()), torch.ones_like(exp_ad)
-    )
-
-    grad_adjoint = torch.autograd.grad(
-        exp_adjoint, tuple(values_adjoint.values()), torch.ones_like(exp_adjoint)
-    )
-
-    assert len(grad_ad) == len(grad_adjoint)
-    for i in range(len(grad_ad)):
-        assert torch.allclose(grad_ad[i], grad_adjoint[i], atol=GRADCHECK_ATOL)
-
-
-@pytest.mark.parametrize("dtype", [torch.complex64, torch.complex128])
-@pytest.mark.parametrize("batch_size", [1, 5])
-@pytest.mark.parametrize("n_qubits", [3, 4])
-def test_differentiate_circuit(
-    dtype: torch.dtype, batch_size: int, n_qubits: int
-) -> None:
-    ops = [
-        pyq.Y(1),
-        pyq.RX(0, "theta_0"),
-        pyq.PHASE(0, "theta_1"),
-        pyq.CSWAP(0, (1, 2)),
-        pyq.CRX(1, 2, "theta_2"),
-        pyq.CPHASE(1, 2, "theta_3"),
-        pyq.CNOT(0, 1),
-        pyq.Toffoli((2, 1), 0),
-    ]
-    theta_0_value = torch.rand(1, dtype=dtype)
-    theta_1_value = torch.rand(1, dtype=dtype)
-    theta_2_value = torch.rand(1, dtype=dtype)
-    theta_3_value = torch.rand(1, dtype=dtype)
-    circ = pyq.QuantumCircuit(n_qubits, ops).to(dtype)
-    state = pyq.random_state(n_qubits, batch_size, dtype=dtype)
-
-    theta_0_ad = torch.tensor([theta_0_value], requires_grad=True)
-    theta_0_adjoint = torch.tensor([theta_0_value], requires_grad=True)
-
-    theta_1_ad = torch.tensor([theta_1_value], requires_grad=True)
-    theta_1_adjoint = torch.tensor([theta_1_value], requires_grad=True)
-
-    theta_2_ad = torch.tensor([theta_2_value], requires_grad=True)
-    theta_2_adjoint = torch.tensor([theta_2_value], requires_grad=True)
-
-    theta_3_ad = torch.tensor([theta_3_value], requires_grad=True)
-    theta_3_adjoint = torch.tensor([theta_3_value], requires_grad=True)
-
-    values_ad = torch.nn.ParameterDict(
-        {
-            "theta_0": theta_0_ad,
-            "theta_1": theta_1_ad,
-            "theta_2": theta_2_ad,
-            "theta_3": theta_3_ad,
-        }
-    ).to(COMPLEX_TO_REAL_DTYPES[dtype])
-    values_adjoint = torch.nn.ParameterDict(
-        {
-            "theta_0": theta_0_adjoint,
-            "theta_1": theta_1_adjoint,
-            "theta_2": theta_2_adjoint,
-            "theta_3": theta_3_adjoint,
-        }
-    ).to(COMPLEX_TO_REAL_DTYPES[dtype])
-
-    obs = pyq.QuantumCircuit(n_qubits, [pyq.Z(0)]).to(dtype)
-    exp_ad = expectation(circ, state, values_ad, obs, DiffMode.AD)
-    exp_adjoint = expectation(circ, state, values_adjoint, obs, DiffMode.ADJOINT)
-
-    grad_ad = torch.autograd.grad(
-        exp_ad, tuple(values_ad.values()), torch.ones_like(exp_ad)
-    )
-
-    grad_adjoint = torch.autograd.grad(
-        exp_adjoint, tuple(values_adjoint.values()), torch.ones_like(exp_adjoint)
-    )
-
-    assert len(grad_ad) == len(grad_adjoint)
-    for i in range(len(grad_ad)):
-        assert torch.allclose(grad_ad[i], grad_adjoint[i], atol=GRADCHECK_ATOL)
 
 
 def test_device_inference() -> None:
@@ -142,25 +23,6 @@ def test_device_inference() -> None:
     circ = pyq.QuantumCircuit(2, ops)
     nested_circ = pyq.QuantumCircuit(2, [circ, circ])
     assert nested_circ._device is not None
-
-
-def test_adjoint_duplicate_params() -> None:
-    n_qubits = 2
-    ops = [pyq.RX(0, param_name="theta_0"), pyq.RX(0, param_name="theta_0")]
-    theta_vals = torch.arange(0, torch.pi, 0.05, requires_grad=True)
-    circ = pyq.QuantumCircuit(n_qubits, ops)
-    obs = pyq.QuantumCircuit(n_qubits, [pyq.Z(0)])
-    init_state = pyq.zero_state(n_qubits)
-    values = {"theta_0": theta_vals}
-    exp_ad = expectation(circ, init_state, values, obs, DiffMode.AD)
-    exp_adjoint = expectation(circ, init_state, values, obs, DiffMode.ADJOINT)
-    grad_ad = torch.autograd.grad(
-        exp_ad, tuple(values.values()), torch.ones_like(exp_ad)
-    )[0]
-    grad_adjoint = torch.autograd.grad(
-        exp_adjoint, tuple(values.values()), torch.ones_like(exp_adjoint)
-    )[0]
-    assert torch.allclose(grad_ad, grad_adjoint, atol=GRADCHECK_ATOL)
 
 
 @pytest.mark.parametrize("fn", [pyq.X, pyq.Z, pyq.Y])
@@ -237,49 +99,6 @@ def test_merge_nested_dict() -> None:
     mergecirc(pyq.random_state(2), vals)
 
 
-@pytest.mark.xfail  # investigate
-@pytest.mark.parametrize("dtype", [torch.complex64, torch.complex128])
-@pytest.mark.parametrize("batch_size", [1, 5])
-@pytest.mark.parametrize("n_qubits", [2])
-def test_adjoint_scale(dtype: torch.dtype, batch_size: int, n_qubits: int) -> None:
-    ops = [pyq.Scale(pyq.X(0), "theta_4")]
-
-    theta_4_value = torch.rand(1, dtype=dtype)
-    circ = pyq.QuantumCircuit(n_qubits, ops).to(dtype)
-
-    state = pyq.random_state(n_qubits, batch_size, dtype=dtype)
-
-    theta_4_ad = torch.tensor([theta_4_value], requires_grad=True)
-    theta_4_adjoint = torch.tensor([theta_4_value], requires_grad=True)
-
-    values_ad = torch.nn.ParameterDict(
-        {
-            "theta_4": theta_4_ad,
-        }
-    ).to(COMPLEX_TO_REAL_DTYPES[dtype])
-    values_adjoint = torch.nn.ParameterDict(
-        {
-            "theta_4": theta_4_adjoint,
-        }
-    ).to(COMPLEX_TO_REAL_DTYPES[dtype])
-
-    obs = pyq.QuantumCircuit(n_qubits, [pyq.Z(0)]).to(dtype)
-    exp_ad = expectation(circ, state, values_ad, obs, DiffMode.AD)
-    exp_adjoint = expectation(circ, state, values_adjoint, obs, DiffMode.ADJOINT)
-
-    grad_ad = torch.autograd.grad(
-        exp_ad, tuple(values_ad.values()), torch.ones_like(exp_ad)
-    )
-
-    grad_adjoint = torch.autograd.grad(
-        exp_adjoint, tuple(values_adjoint.values()), torch.ones_like(exp_adjoint)
-    )
-
-    assert len(grad_ad) == len(grad_adjoint)
-    for i in range(len(grad_ad)):
-        assert torch.allclose(grad_ad[i], grad_adjoint[i], atol=GRADCHECK_ATOL)
-
-
 @pytest.mark.parametrize("n_qubits", [{"low": 2, "high": 5}], indirect=True)
 @pytest.mark.parametrize("batch_size", [{"low": 1, "high": 5}], indirect=True)
 def test_noise_circ(
@@ -315,106 +134,3 @@ def test_sample_run() -> None:
     assert torch.allclose(wf, product_state("1100"))
     assert torch.allclose(pyq.QuantumCircuit(4, [pyq.I(0)]).run("1100"), wf)
     assert "1100" in samples[0]
-
-
-@pytest.mark.parametrize("n_qubits", [3, 4, 5])
-@pytest.mark.parametrize("same_angle", [True, False])
-def test_all_diff_singlegap(n_qubits: int, same_angle: bool) -> None:
-    name_angle_1, name_angle_2 = "theta_0", "theta_1"
-    if same_angle:
-        name_angle_2 = name_angle_1
-
-    ops_rx = pyq.Sequence([pyq.RX(i, param_name=name_angle_1) for i in range(n_qubits)])
-    ops_rz = pyq.Sequence([pyq.RZ(i, param_name=name_angle_2) for i in range(n_qubits)])
-    cnot = pyq.CNOT(1, 2)
-    ops = [ops_rx, ops_rz, cnot]
-
-    circ = pyq.QuantumCircuit(n_qubits, ops)
-    obs = pyq.QuantumCircuit(n_qubits, [pyq.Z(0)])
-
-    theta_0_value = torch.pi / 2
-
-    state = pyq.zero_state(n_qubits)
-
-    theta_0 = torch.tensor([theta_0_value], requires_grad=True)
-
-    if same_angle:
-        values = {name_angle_1: theta_0}
-    else:
-        theta_1_value = torch.pi
-        theta_1 = torch.tensor([theta_1_value], requires_grad=True)
-        values = {name_angle_1: theta_0, name_angle_2: theta_1}
-
-    exp_ad = expectation(circ, state, values, obs, DiffMode.AD)
-    exp_adjoint = expectation(circ, state, values, obs, DiffMode.ADJOINT)
-    exp_gpsr = expectation(circ, state, values, obs, DiffMode.GPSR)
-
-    assert torch.allclose(exp_ad, exp_adjoint)
-    assert torch.allclose(exp_ad, exp_gpsr)
-
-    grad_ad = torch.autograd.grad(
-        exp_ad, tuple(values.values()), torch.ones_like(exp_ad), create_graph=True
-    )[0]
-
-    grad_adjoint = torch.autograd.grad(
-        exp_adjoint,
-        tuple(values.values()),
-        torch.ones_like(exp_adjoint),
-        create_graph=True,
-    )[0]
-
-    grad_gpsr = torch.autograd.grad(
-        exp_gpsr, tuple(values.values()), torch.ones_like(exp_gpsr), create_graph=True
-    )[0]
-
-    gradgrad_ad = torch.autograd.grad(
-        grad_ad, tuple(values.values()), torch.ones_like(grad_ad)
-    )
-
-    # TODO higher order adjoint is not yet supported.
-    # gradgrad_adjoint = torch.autograd.grad(
-    #     grad_adjoint, tuple(values.values()), torch.ones_like(grad_adjoint)
-    # )
-
-    gradgrad_gpsr = torch.autograd.grad(
-        grad_gpsr, tuple(values.values()), torch.ones_like(grad_gpsr)
-    )
-
-    # check first order gradients
-    assert len(grad_ad) == len(grad_adjoint) == len(grad_gpsr)
-    for i in range(len(grad_ad)):
-        assert torch.allclose(
-            grad_ad[i], grad_adjoint[i], atol=GRADCHECK_ATOL
-        ) and torch.allclose(grad_ad[i], grad_gpsr[i], atol=GRADCHECK_ATOL)
-
-    assert len(gradgrad_ad) == len(gradgrad_gpsr)
-
-    # check second order gradients
-    for i in range(len(gradgrad_ad)):
-        assert torch.allclose(gradgrad_ad[i], gradgrad_gpsr[i], atol=GRADCHECK_ATOL)
-
-
-@pytest.mark.xfail(raises=ValueError)
-@pytest.mark.parametrize("gate_type", ["scale", "hamevo"])
-def test_compatibility_gpsr(gate_type: str) -> None:
-
-    if gate_type == "scale":
-        seq_gate = pyq.Sequence([pyq.X(0)])
-        scale = pyq.Scale(seq_gate, "theta_0")
-        ops = [scale]
-    else:
-        hamevo = pyq.HamiltonianEvolution(pyq.Sequence([pyq.X(0)]), "theta_0", (0,))
-
-        ops = [hamevo]
-
-    circ = pyq.QuantumCircuit(1, ops)
-    obs = pyq.QuantumCircuit(1, [pyq.Z(0)])
-    state = pyq.zero_state(1)
-
-    param_value = torch.pi / 2
-    values = {"theta_0": torch.tensor([param_value], requires_grad=True)}
-    exp_gpsr = expectation(circ, state, values, obs, DiffMode.GPSR)
-
-    grad_gpsr = torch.autograd.grad(
-        exp_gpsr, tuple(values.values()), torch.ones_like(exp_gpsr)
-    )
